datapath: Fix ICMP ECHO tuple ports
Previously, when an ICMP EchoRequest was sent from one node to another
with ID > NAT_MIN_EGRESS, the ICMP EchoReply sent from the node to the
outside created a CT entry and NAT entries which could not be related
by GC. E.g. (192.168.34.12 pings 192.168.34.11):

    ICMP IN 192.168.34.12:0 -> 192.168.34.11:38193 XLATE_DST
    192.168.34.11:38193 Created=6292sec HostLocal=1
    ICMP OUT 192.168.34.11:38193 -> 192.168.34.12:0 XLATE_SRC
    192.168.34.11:38193 Created=6292sec HostLocal=1

    ICMP OUT 192.168.34.11:0 -> 192.168.34.12:38193 expires=16783063
    RxPackets=0 RxBytes=0 RxFlagsSeen=0x00 LastRxReport=0 TxPackets=1
    TxBytes=50 TxFlagsSeen=0x00 LastTxReport=16783005 Flags=0x0000 [ ]
    RevNAT=0 SourceSecurityID=0 IfIndex=0

This made the NAT entries to escape the CT GC.

Fix this by setting ICMP ECHO ID in a port which belongs to addr of
the node.

Signed-off-by: Martynas Pumputis <m@lambda.lt>